REDSHOP-4606 Use entity for Product::getCategories
